wayfinding: Move component props from separated file to the component
